// project/src/models/enums/BaseClasses.ts
export enum BaseClasses
    {
    WEAPON = "5422acb9af1c889c16000029",
    UBGL = "55818b014bdc2ddc698b456b",
    ARMOR = "5448e54d4bdc2dcc718b4568",
    ARMOREDEQUIPMENT = "57bef4c42459772e8d35a53b",
    HEADWEAR = "5a341c4086f77401f2541505",
    FACECOVER = "5a341c4686f77469e155819e",
    VEST = "5448e5284bdc2dcb718b4567",
    BACKPACK = "5448e53e4bdc2d60728b4567",
    COMPOUND = "566162e44bdc2d3f298b4573",
    VISORS = "5448e5724bdc2ddf718b4568",
    FOOD = "5448e8d04bdc2ddf718b4569",
    GAS_BLOCK = "56ea9461d2720b67698b456f",
    RAIL_COVER = "55818b1d4bdc2d5b648b4572",
    DRINK = "5448e8d64bdc2dce718b4568",
    BARTER_ITEM = "5448eb774bdc2d0a728b4567",
    INFO = "5448ecbe4bdc2d60728b4568",
    MEDKIT = "5448f39d4bdc2d0a728b4568",
    DRUGS = "5448f3a14bdc2d27728b4569",
    STIMULATOR= "5448f3a64bdc2d60728b456a",
    MEDICAL = "5448f3ac4bdc2dce718b4569",
    MEDICAL_SUPPLIES= "57864c8c245977548867e7f1",
    MOD = "5448fe124bdc2da5018b4567",
    FUNCTIONAL_MOD = "550aa4154bdc2dd8348b456b",
    FUEL = "5d650c3e815116009f6201d2",
    GEAR_MOD= "55802f3e4bdc2de7118b4584",
    STOCK = "55818a594bdc2db9688b456a",
    FOREGRIP = "55818af64bdc2d5b648b4570",
    MASTER_MOD = "55802f4a4bdc2ddb688b4569",
    MOUNT = "55818b224bdc2dde698b456f",
    MUZZLE = "5448fe394bdc2d0d028b456c",
    SIGHTS = "5448fe7a4bdc2d6f028b456b",
    MEDS = "543be5664bdc2dd4348b4569",
    MAP = "567849dd4bdc2d150f8b456e",
    MONEY = "543be5dd4bdc2deb348b4569",
    NIGHTVISION = "5a2c3a9486f774688b05e574",
    KEY = "543be5e94bdc2df1348b4568",
    KEY_MECHANICAL = "5c99f98d86f7745c314214b3",
    KEYCARD = "5c164d2286f774194c5e69fa",
    EQUIPMENT = "543be5f84bdc2dd4348b456a",
    THROW_WEAPON = "543be6564bdc2df4348b4568",
    FOOD_DRINK = "543be6674bdc2df1348b4569",
    PISTOL = "5447b5cf4bdc2d65278b4567",
    REVOLVER = "617f1ef5e8b54b0998387733",
    SMG = "5447b5e04bdc2d62278b4567",
    ASSAULT_RIFLE = "5447b5f14bdc2d61278b4567",
    ASSAULT_CARBINE = "5447b5fc4bdc2d87278b4567",
    SHOTGUN = "5447b6094bdc2dc3278b4567",
    MARKSMAN_RIFLE = "5447b6194bdc2d67278b4567",
    SNIPER_RIFLE = "5447b6254bdc2dc3278b4568",
    MACHINE_GUN ="5447bed64bdc2d97278b4568",
    GRENADE_LAUNCHER= "5447bedf4bdc2d87278b4568",
    SPECIAL_WEAPON= "5447bee84bdc2dc3278b4569",
    SPEC_ITEM = "5447e0e74bdc2d3c308b4567",
    KNIFE = "5447e1d04bdc2dff2f8b4567",
    AMMO = "5485a8684bdc2da71d8b4567",
    AMMO_BOX = "543be5cb4bdc2deb348b4568",
    LOOT_CONTAINER= "566965d44bdc2d814c8b4571",
    MOD_CONTAINER = "5448bf274bdc2dfc2f8b456a",
    SEARCHABLE_ITEM= "566168634bdc2d144c8b456c",
    STASH = "566abbb64bdc2d144c8b457d",
    SORTING_TABLE= "6050cac987d3f925bf016837",
    LOCKABLE_CONTAINER= "5671435f4bdc2d96058b4569",
    SIMPLE_CONTAINER ="5795f317245977243854e041",
    INVENTORY= "55d720f24bdc2d88028b456d",
    STATIONARY_CONTAINER= "567583764bdc2d98058b456e",
    POCKETS = "557596e64bdc2dc2118b4571",
    ARMBAND = "5b3f15d486f77432d0509248",
    DOG_TAG_USEC= "59f32c3b86f77472a31742f0",
    DOG_TAG_BEAR = "59f32bb586f774757e1e8442",
    JEWELRY = "57864a3d24597754843f8721",
    ELECTRONICS = "57864a66245977548f04a81f",
    BUILDING_MATERIAL= "57864ada245977548638de91",
    TOOL= "57864bb7245977548b3b66c2",
    HOUSEHOLD_GOODS= "57864c322459775490116fbf",
    LUBRICANT= "57864e4c24597754843f8723",
    BATTERY= "57864ee62459775490116fc1",
    ASSAULT_SCOPE= "55818add4bdc2d5b648b456f",
    TACTICAL_COMBO= "55818b164bdc2ddc698b456c",
    FLASHLIGHT = "55818b084bdc2d5b648b4571",
    MAGAZINE= "5448bc234bdc2d3c308b4569",
    LIGHT_LASER_DESIGNATOR= "55818b0e4bdc2dde698b456e",
    FLASH_HIDER= "550aa4bf4bdc2dd6348b456b",
    COLLIMATOR= "55818ad54bdc2ddc698b4569",
    IRON_SIGHT= "55818ac54bdc2d5b648b456e",
    COMPACT_COLLIMATOR= "55818acf4bdc2dde698b456b",
    COMPENSATOR= "550aa4af4bdc2dd4348b456e",
    OPTIC_SCOPE= "55818ae44bdc2dde698b456c",
    SPECIAL_SCOPE= "55818aeb4bdc2ddc698b456a",
    OTHER= "590c745b86f7743cc433c5f2",
    SILENCER= "550aa4cd4bdc2dd8348b456c",
    PORTABLE_RANGE_FINDER= "61605ddea09d851a0a0c1bbc",
    ITEM= "54009119af1c881c07000029",
    CYLINDER_MAGAZINE= "610720f290b75a49ff2e5e25"
}